==> tfl-rag-pipeline/01_scraping.ipynb <==
import requests
from bs4 import BeautifulSoup
import regex as re

def access_link(link):
    return requests.get(link,headers={'User-Agent':'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/26.2 Safari/605.1.15'})

response = access_link('https://haveyoursay.tfl.gov.uk/projects')

soup = BeautifulSoup(response.text,'html.parser')


links = soup.find_all('a', class_='project-tile__link')
for tag in links:
    print(tag['href'])

from urllib.parse import urljoin
project_links = {}
for project_link in links:
    link = urljoin('https://haveyoursay.tfl.gov.uk',project_link['href'])
    project_name = str(project_link['href'])[1:]
    project_links[project_name] = link

print(project_links)

import time
import json
import mimetypes

try:
    with open('data/web_data.json', 'r') as f:
        project_data = json.load(f)
    print('data/web_data.json found, skipping web scrape...')
except FileNotFoundError:
    project_data = {}

    items = list(project_links.items())

    for i, item in enumerate(items):
        
        pname = item[0]
        plink = item[1]
        print('Beginning scrape of ' + pname)

        if pname in project_data:
            print(pname + ' found, skipping...')
            continue
        else:
            response = access_link(plink)
            soup = BeautifulSoup(response.text,'html.parser')    

            project_texts = soup.find_all(class_=re.compile('project_widgets')) + soup.find_all(id=re.compile('project_description_text'))
            ptext = ''
            for txt in project_texts:
                ptext += txt.get_text(separator='\n')
            
            project_data[pname] = {}
            project_data[pname]['project_text'] = ptext
            
            documents = soup.find_all('a',class_=re.compile('document'))
            file_number = 0
            project_data[pname]['documents'] = []

            for j, file in enumerate(documents):
                download_link = file['href'] + '/download'

                pdf_response = access_link(download_link)
                print(pdf_response.headers.get('Content-Type'))
                ext = mimetypes.guess_extension(pdf_response.headers.get('Content-Type').split(';')[0])
                if ext == None:
                    print('file extension not found, aborting')
                    break
                document_name = pname + '_' + str(j) + ext
                filepath = 'documents/' + document_name
                with open(filepath,'wb') as f:
                    f.write(pdf_response.content)
                print('Saving file: ' + document_name)
                project_data[pname]['documents'].append(filepath)
            
            time.sleep(1)
        
        with open('web_data.json','w') as f:
            json.dump(project_data,f)
            print(f'Web-scraped data saved under {f}')
