==> tfl-rag-pipeline/02_extraction_and_embedding.ipynb <==
import json
import pdfplumber

try:
    with open('data/web_data.json', 'r') as f:
        project_data = json.load(f)
except FileNotFoundError:
    raise SystemExit('data/web_data.json not found, run the scraping notebook first!')

test_project = project_data['walking-and-cycling-changes-oval-elephant-and-castle']
document_text_pdf = []
for document in test_project['documents']:
    with pdfplumber.open(document) as pdf:
        for page in pdf.pages:
            text = page.extract_text_simple()
            document_text_pdf.append(text)

document_text_pdf[:5]


from docx import Document

test_project = project_data['30-205-bus-proposals']
document_text_docx = []
for document in test_project['documents']:
    if document.split('.')[-1] == 'docx':
        doc = Document(document)
        for paragraph in doc.paragraphs:
            document_text_docx.append(paragraph.text)
        
document_text_docx[:5]

try:
    with open('data/text_data.json', 'r') as f:
            project_data = json.load(f)
except:
    print('no text_data.json found, searching for web_data.json instead...')
    try:
        with open('data/web_data.json', 'r') as f:
            project_data = json.load(f)
    except:
        print('no web_data.json found, run the scraping notebook first!')
        pass



for pname, pdata in project_data.items():
    if 'doc_texts' in pdata and pdata['doc_texts']:
        print(f'doc_texts found for {pname}, skipping...')
        continue
    project_data[pname]['doc_texts'] = []
    for document in pdata['documents']:
        ext = document.split('.')[-1]
        try:
            if ext == 'pdf':
                with pdfplumber.open(document) as pdf:
                    doc_text_pdf = ''
                    for page in pdf.pages:
                        text = page.extract_text_simple()
                        doc_text_pdf += text
                project_data[pname]['doc_texts'].append(doc_text_pdf)
            elif ext == 'docx':
                doc = Document(document)
                text = '\n'.join(p.text for p in doc.paragraphs if p.text.strip())
                project_data[pname]['doc_texts'].append(text)
            print(f'converted {ext} text for {document}')
        except:
            continue

    with open('data/text_data.json','w') as f:
        json.dump(project_data,f)
    
with open('data/text_data.json','w') as f:
    json.dump(project_data,f)
print('text data saved under data/text_data.json')
             


import pandas as pd
import matplotlib.pyplot as plt

pnames = []
doc_names = []
doc_lengths = []
doc_texts = []
for pname, pdata in project_data.items():
    pnames.append(pname)
    doc_names.append(f'{pname}-text')
    doc_texts.append(pdata['project_text'])
    doc_length = len(pdata['project_text'])
    doc_lengths.append(doc_length)
    for i, document in enumerate(pdata['doc_texts']):    
        doc_name = pdata['documents'][i].split('/')[-1]
        pnames.append(pname)
        doc_names.append(doc_name)
        doc_length = len(document.split())
        doc_lengths.append(doc_length)
        doc_texts.append(pdata['doc_texts'][i])
    

doc_df = pd.DataFrame({'project':pnames,'document':doc_names,'doc_length':doc_lengths, 'doc_text':doc_texts})
print(doc_df.head(20))

from langchain_text_splitters import RecursiveCharacterTextSplitter

text_splitter = RecursiveCharacterTextSplitter(chunk_size=500,chunk_overlap=10)

doc_df['doc_chunks'] = doc_df['doc_text'].apply(lambda x: text_splitter.split_text(x))
doc_df['num_chunks'] = doc_df['doc_chunks'].apply(len)

doc_df = doc_df[doc_df['num_chunks'] > 0]
doc_df['num_chunks'].describe()

chunks_df = doc_df.explode('doc_chunks')
chunks_df.head()


from sentence_transformers import SentenceTransformer
import os
import uuid

if os.path.exists('data/embedded_chunks_df.pkl'):
    embedded_chunks_df = pd.read_pickle('data/embedded_chunks_df.pkl')
else:
    model = SentenceTransformer('all-MiniLM-L6-v2')
    embeddings = model.encode(chunks_df['doc_chunks'].tolist(), show_progress_bar=True)
    embedded_chunks_df = chunks_df.copy()
    embedded_chunks_df['embeddings'] = list(embeddings)
    embedded_chunks_df = embedded_chunks_df.drop_duplicates(subset=['doc_chunks'], keep='first')
    embedded_chunks_df['uuid'] = embedded_chunks_df.apply(lambda _: str(uuid.uuid1()), axis=1)

    embedded_chunks_df.to_pickle('data/embedded_chunks_df.pkl')
