==> tfl-rag-pipeline/03_chromadb_rag.ipynb <==
import chromadb
from chromadb.utils.batch_utils import create_batches
import pandas as pd

try:
    embedded_chunks_df = pd.read_pickle('data/embedded_chunks_df.pkl')
except:
    raise SystemExit('data/embedded_chunks_df.pkl not found, run the extraction/embedding notebook first!')

chroma_client = chromadb.PersistentClient()

batches = create_batches(
    api=chroma_client,
    ids=embedded_chunks_df['uuid'].tolist(),
    embeddings = embedded_chunks_df['embeddings'].tolist(),
    documents = embedded_chunks_df['doc_chunks'].tolist(),
    metadatas = [{'document': i} for i in embedded_chunks_df['document'].tolist()])

try:
    chroma_client.delete_collection(name='consultations')
except:
    pass
collection = chroma_client.create_collection(name='consultations')

for batch in batches:
    print(f'adding batch of size {len(batch[0])}')
    collection.add(
        ids = batch[0],
        embeddings = batch[1],
        documents=batch[3],
        metadatas = batch[2]
    )
print('Complete!')



collection.query(query_texts='Oxford Street pedestrianisation?')
