==> tfl-rag-pipeline/04_llm.ipynb <==
import pandas as pd
import ollama
import chromadb

chroma_client = chromadb.PersistentClient()
collection = chroma_client.get_collection(name='consultations')

def run_prompt(query):
    contexts = collection.query(query_texts=query, n_results=10)
    context = ''
    for i, chunk in enumerate(contexts['metadatas'][0]):
        context = context +  f'{chunk["document"]}: {contexts["documents"][0][i]} \n'
    
    prompt = f"""Answer the question based on the context. Provide citations based on the document used. Provide dates where possible.
     
    Question: {query} 
    Context: {context}
    """
    response = ollama.chat(
        model='llama3.1:8b',
        messages=[{'role':'user', 'content': prompt}])

    print(response['message']['content'])

query = input()
run_prompt(query)
